simply functional again to satisfy mypy
