+++ guppylang/src/guppylang/std/qsystem/sol/functional.py
@@ -9,16 +9,19 @@ from typing import no_type_check
 import guppylang.std.qsystem.sol as sol
 from guppylang.decorator import guppy
 from guppylang.std.angles import angle
+from guppylang.std.array import array
 from guppylang.std.builtins import owned
-from guppylang.std.qsystem._functional_base import _make_shared_functional
-from guppylang.std.quantum import qubit
+from guppylang.std.quantum import Measurement, qubit
 
-globals().update(_make_shared_functional(sol))
+N = guppy.nat_var("N")
 
 
-# ---------------------------------------------------------------------------
-# Sol-specific 2-qubit gate
-# ---------------------------------------------------------------------------
+@guppy
+@no_type_check
+def phased_x(q: qubit @ owned, angle1: angle, angle2: angle) -> qubit:
+    """Functional PhasedX gate command."""
+    sol.phased_x(q, angle1, angle2)
+    return q
 
 
 @guppy
@@ -29,3 +32,76 @@ def phased_xx(
     """Functional PhasedXX gate command. The native 2-qubit entangling gate on Sol."""
     sol.phased_xx(q1, q2, angle1, angle2)
     return q1, q2
+
+
+@guppy
+@no_type_check
+def rz(q: qubit @ owned, angle: angle) -> qubit:
+    """Functional Rz gate command."""
+    sol.rz(q, angle)
+    return q
+
+
+@guppy
+@no_type_check
+def measure(q: qubit @ owned) -> Measurement:
+    """Functional measurement command."""
+    return sol.measure(q)
+
+
+@guppy
+@no_type_check
+def measure_and_reset(q: qubit @ owned) -> tuple[qubit, Measurement]:
+    """Functional measure_and_reset command."""
+    b = sol.measure_and_reset(q)
+    return q, b
+
+
+@guppy
+@no_type_check
+def reset(q: qubit @ owned) -> qubit:
+    """Functional Reset command."""
+    sol.reset(q)
+    return q
+
+
+@guppy
+@no_type_check
+def qfree(q: qubit @ owned) -> None:
+    """Functional qfree command."""
+    sol.qfree(q)
+
+
+@guppy
+@no_type_check
+def lazy_measure_and_reset(q: qubit @ owned) -> tuple[qubit, Measurement]:
+    """Functional lazy_measure_and_reset command."""
+    measurement = sol.lazy_measure_and_reset(q)
+    return q, measurement
+
+
+@guppy
+@no_type_check
+def measure_array(qubits: array[qubit, N] @ owned) -> array[Measurement, N]:
+    """Functional measure_array command."""
+    return sol.measure_array(qubits)
+
+
+@guppy
+@no_type_check
+def measure_and_reset_array(
+    qubits: array[qubit, N] @ owned,
+) -> tuple[array[qubit, N], array[Measurement, N]]:
+    """Functional measure_and_reset_array command."""
+    bs = sol.measure_and_reset_array(qubits)
+    return qubits, bs
+
+
+@guppy
+@no_type_check
+def lazy_measure_and_reset_array(
+    qubits: array[qubit, N] @ owned,
+) -> tuple[array[qubit, N], array[Measurement, N]]:
+    """Functional lazy_measure_and_reset_array command."""
+    measurements = sol.lazy_measure_and_reset_array(qubits)
+    return qubits, measurements
